move brat document and annotation types to pie_dataset.document.types
